// backend/src/enums/env.enum.ts
export enum EnvVariablesEnum {
  PORT = 'PORT',
  MONGO_URI = 'MONGO_URI',
  JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRES_IN = 'JWT_EXPIRES_IN',
  ACCESS_TOKEN_EXPIRES_IN = 'ACCESS_TOKEN_EXPIRES_IN',
  REFRESH_TOKEN_EXPIRES_IN = 'REFRESH_TOKEN_EXPIRES_IN',
  MAX_ACTIVE_SESSIONS_PER_USER = 'MAX_ACTIVE_SESSIONS_PER_USER',
  MAX_ACCESS_TOKENS_PER_REFRESH_TOKEN = 'MAX_ACCESS_TOKENS_PER_REFRESH_TOKEN',
  CORS_ALLOWED_ORIGINS = 'CORS_ALLOWED_ORIGINS',
  COOKIE_SECURE = 'COOKIE_SECURE',
  COOKIE_SAME_SITE = 'COOKIE_SAME_SITE',
}