// backend/src/enums/error-messages.enum.ts
export enum ErrorMessageEnum {
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  UNABLE_TO_LOGOUT = 'UNABLE_TO_LOGOUT',
  INVALID_ACCESS_TOKEN = 'INVALID_ACCESS_TOKEN',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  USER_PROFILE_NOT_AVAILABLE = 'USER_PROFILE_NOT_AVAILABLE',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
}